% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_test.R
\name{EstDep}
\alias{EstDep}
\title{Testing independence based on Blest measure between random variables}
\usage{
EstDep(x)
}
\arguments{
\item{x}{Data matrix}
}
\value{
\item{stat}{Blest  statistics from multilinear copula}

\item{pvalue}{P-values for the test statistics}

x <-rpois(100,3)
y <-rpois(100,4)
z <-cbind(x,y)
out<-EstDep(z)
}
\description{
This function computes the matrix of Blest statistics and p-value between random variables with discrete distributions.
}
