% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_dep.R
\name{stat_dep}
\alias{stat_dep}
\title{Computes the Blest measure for tests of independence}
\usage{
stat_dep(x, y)
}
\arguments{
\item{x}{Vector of length n}

\item{y}{Vector of length n}
}
\value{
\item{rho}{Blest measure}
}
\description{
This function Computes the Blest measure for tests of independence used in indep_test
}
\examples{
x <-rpois(100,3)
y <-rpois(100,4)
out <- stat_dep(x,y)

}
\keyword{internal}
